Xcconfig refactor and fix warnings

+++ objc/testcase/benchmark/RepairBenchmark.mm
@@ -31,7 +31,7 @@
     [super setUp];
 
     self.factory.expectedQuality = 100.0; // file size in MB
-    _step = self.factory.expectedQuality * 100;
+    _step = (int) self.factory.expectedQuality * 100;
     self.factory.tolerance = 0.02;
 }
